requirements.txt
requests
pydub
flask
SpeechRecognition
mecab-python3
gensim
python-dotenv